% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_utils.R
\name{create_llm_connection.openai}
\alias{create_llm_connection.openai}
\title{Create LLM connection for OpenAI}
\usage{
\method{create_llm_connection}{openai}(provider, api_key, model, organization_id = NULL, ...)
}
\arguments{
\item{provider}{An object whose class is "openai".}

\item{api_key}{Your OpenAI API key.}

\item{model}{OpenAI model name.}

\item{organization_id}{Optional OpenAI organization ID.}

\item{...}{Additional arguments.}
}
\value{
An "openai_connection" object.
}
\description{
Create LLM connection for OpenAI
}
