% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{describe_plot_object}
\alias{describe_plot_object}
\title{Internal helper to describe a plot object textually}
\usage{
describe_plot_object(plot_obj, plot_name, data = NULL)
}
\arguments{
\item{plot_obj}{The plot object}

\item{plot_name}{Name of the plot in results list}

\item{data}{The original dataset (for context, optional)}
}
\value{
A string describing the plot
}
\description{
Internal helper to describe a plot object textually
}
\keyword{internal}
