% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_utils.R
\name{generate_object_signature_simple}
\alias{generate_object_signature_simple}
\title{Generate a simple signature for statistical objects to detect duplicates}
\usage{
generate_object_signature_simple(object)
}
\arguments{
\item{object}{The statistical object}
}
\value{
A character string signature
}
\description{
Generate a simple signature for statistical objects to detect duplicates
}
\keyword{internal}
