% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_utils.R
\name{send_to_llm.default}
\alias{send_to_llm.default}
\title{Default method for sending prompts to LLMs}
\usage{
\method{send_to_llm}{default}(connection, prompt, ...)
}
\arguments{
\item{connection}{An LLM connection object.}

\item{prompt}{Text prompt to send to the LLM.}

\item{...}{Additional arguments.}
}
\description{
Default method for sending prompts to LLMs
}
