% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_utils.R
\name{send_to_llm.gemini_connection}
\alias{send_to_llm.gemini_connection}
\title{Send prompt to Gemini API}
\usage{
\method{send_to_llm}{gemini_connection}(
  connection,
  prompt,
  data = NULL,
  max_rows = 50,
  max_cols = 20,
  llm_params = list(),
  ...
)
}
\arguments{
\item{connection}{A "gemini_connection" object.}

\item{prompt}{Text prompt to send.}

\item{data}{Optional dataset to include.}

\item{max_rows}{Max rows for including full data.}

\item{max_cols}{Max columns for including full data.}

\item{llm_params}{A list of parameters for Gemini's generationConfig.}

\item{...}{Additional arguments (currently ignored).}
}
\value{
Gemini response text.
}
\description{
Send prompt to Gemini API
}
