% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_utils.R
\name{send_to_llm.openai_connection}
\alias{send_to_llm.openai_connection}
\title{Send prompt to OpenAI API}
\usage{
\method{send_to_llm}{openai_connection}(
  connection,
  prompt,
  data = NULL,
  max_rows = 50,
  max_cols = 20,
  llm_params = list(),
  ...
)
}
\arguments{
\item{connection}{An "openai_connection" object.}

\item{prompt}{Text prompt to send.}

\item{data}{Optional dataset.}

\item{max_rows}{Max rows for including full data.}

\item{max_cols}{Max columns for including full data.}

\item{llm_params}{A list of parameters for OpenAI's API.}

\item{...}{Additional arguments.}
}
\value{
OpenAI response text.
}
\description{
Send prompt to OpenAI API
}
