% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retry_utils.R
\name{set_auto_stat_verbose}
\alias{set_auto_stat_verbose}
\title{Set auto_stat verbosity}
\usage{
set_auto_stat_verbose(verbose = TRUE)
}
\arguments{
\item{verbose}{Logical indicating whether to show detailed messages}
}
\description{
Set auto_stat verbosity
}
